# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Compliance internal api",
    "description": "Compliance internal api.",
    "version": "1.0.0"
  },
  "paths": {
    "/internal/api/individual/referenceNumber/{referenceNumber}": {
      "get": {
        "summary": "Returns an individual",
        "description": "Returns individual profile based on reference number",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Individual response"
          }
        }
      },
      "parameters": [
        {
          "name": "referenceNumber",
          "in": "path",
          "description": "Individual reference number",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }
      ]
    },
    "/internal/api/individual/save/individual": {
      "post": {
        "summary": "Saves individual",
        "description": "Create new / update exisitng individual profile",
        "produces": [
          "text/plain"
        ],
        "responses": {
          "200": {
            "description": "Individual reference number"
          }
        },
        "requestBody": {
          "description": "Individual paylod",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "address"
                ],
                "additionalProperties": false,
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/entity": {
      "post": {
        "summary": "summary todo",
        "description": "description todo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "response description todo"
          }
        },
        "requestBody": {
          "description": "request body todo",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "number"
                  },
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "name",
                  "address"
                ],
                "additionalProperties": false,
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/entity/{id}": {
      "get": {
        "summary": "summary todo",
        "description": "description todo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "response description todo"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    }
  }
}